Clean Up

+++ Commission.cpp
@@ -1,6 +1,6 @@
 
 #include "Commission.h"
 
-int Commission(int nLocks, int nStocks, int nBarrels){
+float Commission(int nLocks, int nStocks, int nBarrels){
 	return Invalid;
 }
